pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "optimized-memory-mcp-serverv2"
version = "0.1.0"
description = "A Python-based Model Context Protocol (MCP) server implementation"
requires-python = ">=3.13.1"
dependencies = [
    "mcp-sdk>=1.0.0",
    "SQLAlchemy>=2.0.0",
    "alembic>=1.13.0",
    "uvx>=0.1.0"
]

[project.scripts]
mcp-server = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --cov=src"
testpaths = ["tests"]
